requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
aiohttp==3.8.4
    # via
    #   aiohttp-retry
    #   langchain
    #   openai
    #   twilio
aiohttp-retry==2.8.3
    # via twilio
aiosignal==1.3.1
    # via aiohttp
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
billiard==3.6.4.0
    # via celery
boto3==1.26.115
    # via -r requirements/requirements.in
botocore==1.29.115
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via whitenoise
celery[redis]==5.2.7
    # via -r requirements/requirements.in
celery-progress==0.3
    # via -r requirements/requirements.in
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==40.0.2
    # via pyjwt
dataclasses-json==0.5.7
    # via langchain
defusedxml==0.7.1
    # via python3-openid
django==4.2
    # via
    #   -r requirements/requirements.in
    #   django-allauth
    #   django-allauth-2fa
    #   django-anymail
    #   django-hijack
    #   django-otp
    #   django-storages
    #   djangorestframework
    #   drf-spectacular
django-allauth==0.54.0
    # via
    #   -r requirements/requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.10.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==9.1
    # via -r requirements/requirements.in
django-environ==0.10.0
    # via -r requirements/requirements.in
django-hijack==3.3.0
    # via -r requirements/requirements.in
django-otp==1.1.6
    # via django-allauth-2fa
django-storages==1.13.2
    # via -r requirements/requirements.in
django-waffle==3.0.0
    # via -r requirements/requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   drf-spectacular
djangorestframework-api-key==2.3.0
    # via -r requirements/requirements.in
drf-spectacular==0.26.2
    # via -r requirements/requirements.in
ffmpeg==1.4
    # via -r requirements/requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via
    #   requests
    #   yarl
inflection==0.5.1
    # via drf-spectacular
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.17.3
    # via drf-spectacular
kombu==5.2.4
    # via celery
langchain==0.0.260
    # via -r requirements/requirements.in
langsmith==0.0.20
    # via langchain
markdown==3.4.3
    # via -r requirements/requirements.in
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.5
    # via langchain
numpy==1.24.2
    # via
    #   langchain
    #   numexpr
oauthlib==3.2.2
    # via requests-oauthlib
openai==0.27.4
    # via -r requirements/requirements.in
openapi-schema-pydantic==1.2.4
    # via langchain
packaging==23.1
    # via marshmallow
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pydantic==1.10.7
    # via
    #   -r requirements/requirements.in
    #   langchain
    #   langsmith
    #   openapi-schema-pydantic
pydub==0.25.1
    # via -r requirements/requirements.in
pyjwt[crypto]==2.6.0
    # via
    #   django-allauth
    #   twilio
pypng==0.20220715.0
    # via qrcode
pyrsistent==0.19.3
    # via jsonschema
pytelegrambotapi==4.12.0
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via botocore
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via
    #   celery
    #   djangorestframework
    #   twilio
pyyaml==6.0
    # via
    #   drf-spectacular
    #   langchain
qrcode==7.4.2
    # via django-allauth-2fa
redis==4.5.4
    # via celery
requests==2.28.2
    # via
    #   django-allauth
    #   django-anymail
    #   langchain
    #   langsmith
    #   openai
    #   pytelegrambotapi
    #   requests-oauthlib
    #   twilio
requests-oauthlib==1.3.1
    # via django-allauth
s3transfer==0.6.0
    # via boto3
sentry-sdk==1.19.1
    # via -r requirements/requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlalchemy==1.4.47
    # via langchain
sqlparse==0.4.4
    # via django
tenacity==8.2.2
    # via langchain
tqdm==4.65.0
    # via openai
twilio==8.5.0
    # via -r requirements/requirements.in
typing-extensions==4.5.0
    # via
    #   pydantic
    #   qrcode
    #   typing-inspect
typing-inspect==0.8.0
    # via dataclasses-json
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.15
    # via
    #   botocore
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
whitenoise[brotli]==6.4.0
    # via -r requirements/requirements.in
yarl==1.8.2
    # via aiohttp
